+++ src/connector/src/source/nexmark/source/generator.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 PieDb Data
+// Copyright 2022 Piedb Data
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -60,7 +60,8 @@ impl NexmarkEventGenerator {
 
             if let Some(event) = last_event.take() {
                 msgs.push(
-                    NexmarkMessage::new(self.split_id.clone(), self.events_so_far, event).into(),
+                    NexmarkMessage::new(self.split_id.clone(), self.events_so_far as u64, event)
+                        .into(),
                 );
             }
 
@@ -100,7 +101,8 @@ impl NexmarkEventGenerator {
                 }
 
                 msgs.push(
-                    NexmarkMessage::new(self.split_id.clone(), self.events_so_far, event).into(),
+                    NexmarkMessage::new(self.split_id.clone(), self.events_so_far as u64, event)
+                        .into(),
                 );
             }
 
@@ -112,7 +114,7 @@ impl NexmarkEventGenerator {
 
             if !self.use_real_time && self.min_event_gap_in_ns > 0 {
                 tokio::time::sleep(Duration::from_nanos(
-                    (self.events_so_far - old_events_so_far) * self.min_event_gap_in_ns,
+                    (self.events_so_far - old_events_so_far) as u64 * self.min_event_gap_in_ns,
                 ))
                 .await;
             }
