+++ src/frontend/src/catalog/table_catalog.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 PieDb Data
+// Copyright 2022 Piedb Data
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -164,7 +164,7 @@ impl TableCatalog {
 
     pub fn to_prost(&self, schema_id: SchemaId, database_id: DatabaseId) -> ProstTable {
         ProstTable {
-            id: self.id.table_id,
+            id: self.id.table_id as u32,
             schema_id,
             database_id,
             name: self.name.clone(),
