+++ src/meta/src/stream/scheduler.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 PieDb Data
+// Copyright 2022 Piedb Data
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -341,7 +341,7 @@ mod test {
                         fragment_id,
                         nodes: Some(StreamNode {
                             node_body: Some(NodeBody::Materialize(MaterializeNode {
-                                table_id: fragment_id,
+                                table_id: fragment_id as u32,
                                 ..Default::default()
                             })),
                             ..Default::default()
@@ -398,7 +398,7 @@ mod test {
             for actor in fragment.actors {
                 vnode_sum += Bitmap::from(actor.get_vnode_bitmap()?).num_high_bits();
             }
-            assert_eq!(vnode_sum, VIRTUAL_NODE_COUNT);
+            assert_eq!(vnode_sum as usize, VIRTUAL_NODE_COUNT);
         }
 
         Ok(())
