+++ src/storage/src/hummock/sstable_store.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 PieDb Data
+// Copyright 2022 Piedb Data
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -607,7 +607,7 @@ impl SstableWriter for StreamingUploadWriter {
         let join_handle = tokio::spawn(async move {
             let uploader_memory_usage = self.object_uploader.get_memory_usage();
             let _tracker = self.tracker.map(|mut t| {
-                    if !t.try_increase_memory(uploader_memory_usage) {
+                    if !t.try_increase_memory(uploader_memory_usage as u64) {
                         tracing::debug!("failed to allocate increase memory for data file, sst id: {}, file size: {}",
                                         self.sst_id, uploader_memory_usage);
                     }
